# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Library Management API",
    "description": "API for managing a collection of books in a library"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/books": {
      "post": {
        "summary": "Add a new book",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Book created"
          }
        }
      },
      "get": {
        "summary": "List all books",
        "responses": {
          "200": {
            "description": "A list of books",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Book"
              }
            }
          }
        }
      }
    },
    "/books/search": {
      "get": {
        "summary": "Search for books",
        "parameters": [
          {
            "name": "author",
            "in": "query",
            "type": "string"
          },
          {
            "name": "published_year",
            "in": "query",
            "type": "integer"
          },
          {
            "name": "genre",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of books",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Book"
              }
            }
          }
        }
      }
    },
    "/books/{isbn}": {
      "delete": {
        "summary": "Delete a book by ISBN",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Book deleted"
          }
        }
      },
      "put": {
        "summary": "Update book details by ISBN",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book updated"
          }
        }
      }
    }
  },
  "definitions": {
    "Book": {
      "type": "object",
      "required": ["title", "author", "published_year", "isbn"],
      "properties": {
        "title": {
          "type": "string"
        },
        "author": {
          "type": "string"
        },
        "published_year": {
          "type": "integer"
        },
        "isbn": {
          "type": "string"
        },
        "genre": {
          "type": "string"
        }
      }
    }
  }
}